TEST run license fallback without Hub
